requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.13
sentence-transformers>=2.2.0
transformers>=4.35.0
torch>=2.1.0
python-dotenv>=1.0.0
pinecone>=3.0.0
pypdf>=3.17.0
numpy>=1.24.0,<2.0.0  # Force NumPy 1.x
pandas>=2.0.0
scikit-learn>=1.3.0
langchain-huggingface
langchain-pinecone

-e .